backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
openai==1.59.9
python-dotenv==1.0.1
